Update `@replayio/playwright` (#353)
This one makes sure that the `@replayio/reply` (transitive dependency)
gets updated too (u can see the change in the lockfile).

With this one, I can record this repo and I get recordings. Some fail at
times when the metadata's size gets exceeded. We know about the issue
and we are in the process of figuring out the solution.