Use a nicer `list` reporter locally (#356)
This addresses the complaint from
https://github.com/bvaughn/react-resizable-panels/pull/348#issuecomment-2102774377
. I didn't realize that the default is `'list'` locally, I added
`'line'` reporter since I've seen that being used in a couple of other
projects. I totally agree that `'list'` is quite nice and I prefer it to
`'line'` - especially if we consider smaller test suites.

⚠️ `@replayio/playwright` still spits out a bunch of junk into the
stdout. That makes the current output very cluttered - even with this
improvement. This is going to be addressed with PRO-315