feat(Support SVG elements on modals): expand utils for stacking order to include SVGElements (#427)
The issue:
When hovering over SVG icons on elements "above" a handle (e.g. modal)
the resize cursor appears and users can drag it to resize the panels
"bellow".

Suggested solution:
Expand the `recalculateIntersectingHandles` function to include
`SVGElement` when looking at the stacking order of elements which fire
pointer events.