Replace innerHtml with insertRule (#444)
The use of `.innerHTML` causes issues for hosts that have opted in to
[Trusted
Types](https://developer.mozilla.org/en-US/docs/Web/API/Trusted_Types_API).
Because `.innerHTML` can act as a vector for DOM-XSS attacks, websites
might lock it down with Trusted Types, causing the specific line to
throw a runtime error.
 
In the case of this specific library, the `.innerHTML` call could just
as easily be replaced with inserting a CSS rule, as follows:
```
styleElement.sheet?.insertRule(`*{cursor: ${style}!important;}`)
```
 
Note that to satisfy TypeScript, null-coalescing is added to the
`.sheet` access. However, when a style element is attached to the DOM,
that `.sheet` property should never be undefined.

Co-authored-by: Subi Kim <subikim@college.harvard.edu>