Attach "pointerup" and "pointercancel" listeners to ownerDocument body (#475)
This PR changes which element the handlePointerUp event handler is added
to, it now adds the event handler to the body element of each owner
docment where a panelResizeHandle is rendered. This accounts for use
cases where the panelResizeHandle is rendered in a document that is in a
child window.

This PR also adds a test for checking which document body the event
handler is added to, which demonstrates a usecase where
PanelResizeHandles are rendered to a separate document.

addresses #468

---------

Co-authored-by: Brian Vaughn <brian.david.vaughn@gmail.com>